+++ R/repair.R
@@ -2,7 +2,7 @@
 # NOTE: mostly for internal use when i accidentally
 # link packages in the user library into the cache
 renv_repair <- function(library = NULL) {
-  library <- library %||% renv_libpaths_default()
+  library <- renv_path_normalize(library %||% renv_libpaths_default())
   paths <- list.files(library, full.names = TRUE)
   links <- Sys.readlink(paths)
   broken <- nzchar(links) & !file.exists(links)