+++ R/restore.R
@@ -71,7 +71,7 @@ restore <- function(project  = NULL,
   project  <- renv_project_resolve(project)
   renv_scope_lock(project = project)
 
-  library  <- library %||% renv_libpaths_all()
+  library  <- renv_path_normalize(library %||% renv_libpaths_all())
   lockfile <- lockfile %||% renv_lockfile_load(project = project)
   lockfile <- renv_lockfile_resolve(lockfile)