+++ R/update.R
@@ -164,7 +164,7 @@ update <- function(packages = NULL,
   renv_scope_lock(project = project)
 
   # get package records
-  library <- library %||% renv_libpaths_all()
+  library <- renv_path_normalize(library %||% renv_libpaths_all())
   records <- renv_snapshot_r_packages(library = library, project = project)
   packages <- packages %||% setdiff(names(records), settings$ignored.packages(project = project))